==> Code/01_Creating_MoH_Dataset.ipynb <==
import csv
import pandas as pd
import numpy as np

#Import LegislatoR datasets 

House_Legislators = pd.read_csv('~/Documents/GitHub/CampaignDonations/Data/Congress_Legislators/LegislatoR_Data/house_us_core.csv')
#House_Legislators.head()  

Political_Legislators = pd.read_csv('~/Documents/GitHub/CampaignDonations/Data/Congress_Legislators/LegislatoR_Data/house_us_political.csv')
#Political_Legislators.head()

# Merge both datasets 
MoH_Legislators = Political_Legislators.join(House_Legislators.set_index('pageid'), on='pageid')
#list(MoH_Legislators.columns)

#To select rows whose column value is in list 
sessions = [113, 114, 115, 116] # Session 117 is not included into the dataset 
MoH_113_116 = MoH_Legislators[MoH_Legislators.session.isin(sessions)]
#list(MoH_113_116.columns)

# Filter out those MoH that only served one term 

Pageid = MoH_113_116
Pageid['counting'] = 1

import numpy as np
df_grouped = Pageid.groupby(["pageid"]).agg(
    count_col=pd.NamedAgg(column="counting", aggfunc="count")
)

#list(Pageid.columns)
#list(MoH_113_116.columns)


# Checking whether it worked
#print(df_grouped)
#df_grouped.sort_values(by=["count_col"])

# Rejoing with main dataframe: left join
MoH_113_116 = pd.merge(MoH_113_116, df_grouped, on='pageid', how='left')
#MoH_113_117.head()


# Drop all MoH that served less than 2 sessions
MoH_113_116 = MoH_113_116[MoH_113_116["count_col"] >= 2]
MoH_113_116 = MoH_113_116.drop(columns=['counting'])
MoH_113_116.head()

# MoH_113_117["count_col"].dtype

MoH_113_116.columns = MoH_113_116.columns.str.replace('count_col', 'Number_Sessions_Served')
#list(MoH_113_116.columns)

# Creating dummy columns for each session in separate datasets
MoH_113 = MoH_113_116[MoH_113_116["session"] == 113]
MoH_113 = MoH_113[["pageid"]]
MoH_113["Session_113"] = True 

MoH_114 = MoH_113_116[MoH_113_116["session"] == 114]
MoH_114 = MoH_114[["pageid"]]
MoH_114["Session_114"] = True 

MoH_115 = MoH_113_116[MoH_113_116["session"] == 115]
MoH_115 = MoH_115[["pageid"]]
MoH_115["Session_115"] = True 

MoH_116 = MoH_113_116[MoH_113_116["session"] == 116]
MoH_116 = MoH_116[["pageid"]]
MoH_116["Session_116"] = True 


# Rejoing session-datasets with main dataframe
MoH_113_116 = pd.merge(MoH_113_116, MoH_113, on='pageid', how='left')

MoH_113_116 = pd.merge(MoH_113_116, MoH_114, on='pageid', how='left')

MoH_113_116 = pd.merge(MoH_113_116, MoH_115, on='pageid', how='left')

MoH_113_116 = pd.merge(MoH_113_116, MoH_116, on='pageid', how='left')


# Drop all column that are session specific so we can just keep one unique row for every MoH

MoH_113_116_Unique = MoH_113_116.drop(columns=['session','constituency','session_start', 'session_end','service','house_majority','house_speaker', 'house_assistant_to_speaker','house_republican_conference_chair','house_republican_conference_vice_chair','house_conference_secretary','house_democratic_caucus_chairman','house_democratic_caucus_vice_chairman','house_steering_committee_chair','house_policy_committee_chair','house_republican_policy_committee_chair','house_campaign_committee_deputy_chair','house_organization_study_review_chair','house_caucus_secretary','house_democratic_campaign_committee_chair','house_democratic_policy_communications_committee_chair','house_democratic_policy_communications_committee_cochairs','house_democratic_steering_and_policy_committee_cochairs','house_democratic_org_study_review_chair','house_committee_on_ways_means_chair','house_republican_campaign_committee_chair','house_majority_leader','house_majority_whip','house_assistant_majority_whip','house_senior_chief_deputy_whip','house_senior_deputy_whip','house_deputy_whip_team','house_assistant_deputy_whip_team','house_majority_chief_deputy_whip','house_minority_leader','house_assistant_minority_leader','house_minority_whip','house_minority_senior_chief_deputy_whip','house_minority_deputy_whip_team','house_minority_chief_deputy_whip','country'])
#list(MoH_113_116.columns)

MoH_113_116['pageid'].nunique() #500 unique pageid in dataset 

length1=len(MoH_113_116_Unique)

MoH_113_116_Unique = MoH_113_116_Unique.drop_duplicates(subset='pageid', keep='first')

# length after removing duplicates
length2=len(MoH_113_116_Unique)
  
# printing all data frame lengths 
print(length1, length2)

# Export Dataframe into CSV
#MoH_113_116_Unique.to_csv(index=False)

compression_opts = dict(method='zip',
                        archive_name='All_MoH_Dataframe.csv')  
MoH_113_116_Unique.to_csv('All_MoH_Dataframe.zip', index=False,
          compression=compression_opts)
