==> Code/03b_Merge_Donations_MoH(Bloomberg_Citadel).ipynb <==
import numpy as np
import pandas as pd



xx = "~/Documents/GitHub/CampaignDonations/Output_Data/All_MoH_Dataframe.csv"

all_moh_1218 = pd.read_csv(xx)
all_moh_1218.head()



all_moh_1218.shape

def sep_first(df):
    # new data frame with split value columns
    data_names = df["name"].str.split(" ", expand = True).copy()
    frames = [df, data_names]
    df = pd.concat(frames, axis = 1)
    df = column_rename(df,0,"first")
    df = column_rename(df,1,"second")
    df = column_rename(df,2,"third")
    df = column_rename(df,3,"fourth")
    return df

def column_rename(df,old,new):
    df = df.rename(columns = {old: new})
    return df

"""All the code to run the cleaning"""
all_moh_1218 = sep_first(all_moh_1218)
first = all_moh_1218['first'].tolist()
second = all_moh_1218['second'].tolist()
third = all_moh_1218['third'].tolist()
fourth = all_moh_1218['fourth'].tolist()
combines = []
position = 0
for item in range(0,len(first)):
    combines.append([first[position],second[position],third[position],fourth[position]])
    position+=1
for item in combines:
    print (item)
# value value none none
# value wrong value none
# value wrong value wrong
len(combines)
name = []

for item in combines:
    if item[2] == None and item[3] == None:
        name.append([item[0], item[1]])
    if item[2] != None and item[3]== None:
        if "Jr." not in item[2]:
            name.append([item[0],item[2]])
        else:
            name.append([item[0],item[1]])
    if item[2] != None and item[3]!= None:
        if "F." in item[0]:
            name.append([item[1],item[2]])
        else:
            name.append([item[0],item[3]])

firstname = []
lastname = []
num = 0
for item in name:
    firstname.append(item[0])
    lastname.append(item[1])
    
all_moh_1218['first'] = firstname
all_moh_1218['second'] = lastname
all_moh_1218 = column_rename(all_moh_1218,"second", 'last')
all_moh_1218['last'] = all_moh_1218['last'].str.replace(',', '')
all_moh_1218 = all_moh_1218.drop(['third', 'fourth'], axis = 1)


all_moh_1218.head()

# Load Bloomberg and Citadel Data 

cc = "~/Documents/GitHub/CampaignDonations/Output_Data/Donation_Data_Bloomberg.csv"
dd = "~/Documents/GitHub/CampaignDonations/Output_Data/Donation_Data_Citadel.csv"


Bloomberg = pd.read_csv(cc)
Citadel = pd.read_csv(cc)

# Change names of donation columns 
Bloomberg = column_rename(Bloomberg, 'donations2012', 'Bloomberg_donations2012')
Bloomberg = column_rename(Bloomberg, 'donations2014', 'Bloomberg_donations2014')
Bloomberg = column_rename(Bloomberg, 'donations2016', 'Bloomberg_donations2016')
Bloomberg = column_rename(Bloomberg, 'donations2018', 'Bloomberg_donations2018')

Citadel = column_rename(Citadel, 'donations2012', 'Citadel_donations2012')
Citadel = column_rename(Citadel, 'donations2014', 'Citadel_donations2014')
Citadel = column_rename(Citadel, 'donations2016', 'Citadel_donations2016')
Citadel = column_rename(Citadel, 'donations2018', 'Citadel_donations2018')




# Create an ID to be able to merge with MOH DAta
Bloomberg["ID_NameParty"] = Bloomberg["first"]+ Bloomberg["last"] + Bloomberg["party"]
Bloomberg = Bloomberg.drop(columns  = ["first", "last"])

Bloomberg.head()

# Create an ID to be able to merge with MOH DAta
Citadel["ID_NameParty"] = Citadel["first"]+ Citadel["last"] + Citadel["party"]
Citadel =Citadel.drop(columns  = ["first", "last"])
Citadel.head()

# Create an ID to be able to merge with Donation DAta
all_moh_1218["ID_NameParty"] = all_moh_1218["first"]+ all_moh_1218["last"] + all_moh_1218["party"]

all_moh_1218.head()


MoH_Vote_Data = pd.merge(all_moh_1218 , Bloomberg , on = "ID_NameParty", how='left')

MoH_Vote_Data = pd.merge(MoH_Vote_Data , Citadel , on = "ID_NameParty", how='left')



sample = MoH_Vote_Data.loc[MoH_Vote_Data['Number_Sessions_Served'] == 4.0]

sample.to_csv("Donation_MoH_Bloomberg_Citadel_NEW.csv",index = False)


test_merge = pd.merge(Bloomberg , all_moh_1218, how="outer")
test_merge.head()

test_merge.shape

test_merge = pd.merge(Citadel, test_merge, how="outer")
test_merge.head()


sample = test_merge.loc[test_merge['Number_Sessions_Served'] == 4.0]

sample.to_csv("Donation_MoH_Bloomberg_Citadel.csv",index = False)

#sample.reset_index()
