==> Code/05_Merge_DonorVoteData_AllCategories.ipynb <==
import csv
import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')


# Merging Voting Data from Different Categories 

# Load the data

a1 = "~/Documents/GitHub/CampaignDonations/Output_Data/02a_02f_Data_Output/MoHs_VoteRatios_113_116_FS_IIR.csv"
a2 = "~/Documents/GitHub/CampaignDonations/Output_Data/02a_02f_Data_Output/MoHs_VoteRatios_113_116_FS_FSS.csv"
a3 = "~/Documents/GitHub/CampaignDonations/Output_Data/02a_02f_Data_Output/MoHs_VoteRatios_113_116_FS_CM.csv"
a4 = "~/Documents/GitHub/CampaignDonations/Output_Data/02a_02f_Data_Output/MoHs_VoteRatios_113_116_FS_REB.csv"

df_Vote_FS_IIR = pd.read_csv(a1)
df_Vote_FS_FSS = pd.read_csv(a2)
df_Vote_FS_CM = pd.read_csv(a3)
df_Vote_FS_REB = pd.read_csv(a4)

df_Vote_FS_FSS = df_Vote_FS_FSS.drop(columns="person")
df_Vote_FS_CM = df_Vote_FS_CM.drop(columns="person")
df_Vote_FS_REB = df_Vote_FS_REB.drop(columns="person")


# Merge 

df_Vote = pd.merge(df_Vote_FS_IIR ,df_Vote_FS_FSS ,on = "name", how='left')
df_Vote = pd.merge(df_Vote  , df_Vote_FS_CM ,on = "name", how='left')
df_Vote = pd.merge(df_Vote  , df_Vote_FS_REB ,on = "name", how='left')

df_Vote.head()

df_Vote["party_short"] = df_Vote["name"]

import re #Regular Expressions

for x in range(len(df_Vote)):
    df_Vote["party_short"].iloc[x] = re.sub(".+D,.+", "D",  df_Vote["party_short"].iloc[x])
    df_Vote["party_short"].iloc[x] = re.sub(".+R,.+", "R",  df_Vote["party_short"].iloc[x])
    
# Setting all non-republicans and democrats to Neutral "N"    
for x in range(len(df_Vote)):
    if (df_Vote["party_short"].iloc[x] != "D" and df_Vote["party_short"].iloc[x] != "R" ):
        df_Vote["party_short"].iloc[x] = "N"    
    



df_Vote["Pure_name"] =  df_Vote["name"]

for x in range(len(df_Vote)):
    df_Vote["Pure_name"].iloc[x] = re.sub("Rep.", "",  df_Vote["Pure_name"].iloc[x])
    df_Vote["Pure_name"].iloc[x] = re.sub("]", "",  df_Vote["Pure_name"].iloc[x])
    df_Vote["Pure_name"].iloc[x] = re.sub("[0-9]", "",  df_Vote["Pure_name"].iloc[x])
    df_Vote["Pure_name"].iloc[x] = re.sub("\[D, -", "",  df_Vote["Pure_name"].iloc[x])
    df_Vote["Pure_name"].iloc[x] = re.sub("\[R, -", "",  df_Vote["Pure_name"].iloc[x])
    df_Vote["Pure_name"].iloc[x] = re.sub("Jr.", "",  df_Vote["Pure_name"].iloc[x])
    
#Problem: By deleting the add-on "Jr." there might some MoH that end up having the same ID 

df_Vote["LastName"] = df_Vote["Pure_name"].str.split().str[-1]
df_Vote["FirstName"] = df_Vote["Pure_name"].str.split().str[0]

# Create an ID to be able to merge with MOH DAta
df_Vote["ID_NameParty"] = df_Vote["FirstName"]+ df_Vote["LastName"] + df_Vote["party_short"]


df_Vote.head()        




#len(df_Vote) #697
unique_MoHs = df_Vote["ID_NameParty"]
length_all = len(unique_MoHs)

unique_MoHs = unique_MoHs.drop_duplicates()

unique_MoHs = pd.DataFrame(unique_MoHs)

length_unique = len(unique_MoHs)

print(length_all , length_unique)


# EXKURS
# Merge with Citadel-Bloomberg Data


# Get Data From other Donors: Bloomberg and Citadel to merge it as well

a1 = "Donation_MoH_Bloomberg_Citadel_NEW.csv"
df_BloomCitadel = pd.read_csv(a1)


df_BloomCitadel.head()


# EXKURS
# Merge datasets 

MoH_Vote_Data = pd.merge(df_BloomCitadel, df_Vote, on = "ID_NameParty", how='left')

MoH_Vote_Data.shape


# EXKURS
# Merge with Vote Data in Category Film (from former disney trial)

# Get the data
a = "~/Documents/GitHub/CampaignDonations/Output_Data/MoHs_VoteRatios_113_116_Film.csv" 
df_FilmVote = pd.read_csv(a)


# Change df_FilmVote a little bit to make sure the already included vote data is properly marked 
df_FilmVote["VoteAlong113thRatio_Film"] = df_FilmVote["VoteAlong113thRatio"] 
df_FilmVote["VoteAlong114thRatio_Film"] = df_FilmVote["VoteAlong114thRatio"] 
df_FilmVote["VoteAlong115thRatio_Film"] = df_FilmVote["VoteAlong115thRatio"] 
df_FilmVote["VoteAlong116thRatio_Film"] = df_FilmVote["VoteAlong116thRatio"] 

list = ["VoteAlong113thRatio", "VoteAlong114thRatio", "VoteAlong115thRatio", "VoteAlong116thRatio"]
df_FilmVote = df_FilmVote.drop(columns = list)



# EXKURS
MoH_Vote_Data2 = pd.merge(MoH_Vote_Data, df_FilmVote, on = "person", how ='left')


# EXKURS

# Save Data-Set externally

compression_opts = dict(method='zip',
                        archive_name='MoH_Vote_AllDonors_Data_New.csv')  
MoH_Vote_Data2.to_csv('MoH_Vote_AllDonors_Data_New.zip', index=False,
          compression=compression_opts)  


#Load DataSet that contains all MoH and the Disney Donation Data

bb = "~/Documents/GitHub/CampaignDonations/Output_Data/walt_disney_moh20122018.csv"

df_MoH = pd.read_csv(bb)

df_MoH.head()

# Change df_MoH a little bit to make sure the already included donations data is properly marked as Dinsey 

df_MoH["disney_donations2012"] = df_MoH["donations2012"] 
df_MoH["disney_donations2014"] = df_MoH["donations2014"] 
df_MoH["disney_donations2016"] = df_MoH["donations2016"] 
df_MoH["disney_donations2018"] = df_MoH["donations2018"] 

list = ["donations2012", "donations2014" ,"donations2016" ,"donations2018"]
df_MoH = df_MoH.drop(columns = list)



# Create an ID in the MoH data to be able to merge with vote data 
df_MoH["ID_NameParty"] = df_MoH["first"] + df_MoH["last"] + df_MoH["party"]

df_MoH.head()

# Get Data From other Donors: Bloomberg and Citadel to merge it as well

a1 = "~/Documents/GitHub/CampaignDonations/Output_Data/Donation_MoH_Bloomberg_Citadel.csv"
df_BloomCitadel = pd.read_csv(a1)

# Create ID_NameParty Column for merging 
df_BloomCitadel["ID_NameParty"] = df_BloomCitadel["first"] + df_BloomCitadel["last"] + df_BloomCitadel["party"]

df_BloomCitadel.head()



# Merge ALl datasets 

MoH_Vote_Data = pd.merge(df_Vote , df_MoH, on = "ID_NameParty", how='left')

MoH_Vote_Data = pd.merge(MoH_Vote_Data  , df_BloomCitadel, on = "ID_NameParty", how='left')

MoH_Vote_Data.head()



# Merge with Vote Data in Category Film (from former disney trial)

# Get the data
a = "~/Documents/GitHub/CampaignDonations/Output_Data/MoHs_VoteRatios_113_116_Film.csv" 
df_FilmVote = pd.read_csv(a)


# Change df_FilmVote a little bit to make sure the already included vote data is properly marked 
df_FilmVote["VoteAlong113thRatio_Film"] = df_FilmVote["VoteAlong113thRatio"] 
df_FilmVote["VoteAlong114thRatio_Film"] = df_FilmVote["VoteAlong114thRatio"] 
df_FilmVote["VoteAlong115thRatio_Film"] = df_FilmVote["VoteAlong115thRatio"] 
df_FilmVote["VoteAlong116thRatio_Film"] = df_FilmVote["VoteAlong116thRatio"] 

list = ["VoteAlong113thRatio", "VoteAlong114thRatio", "VoteAlong115thRatio", "VoteAlong116thRatio"]
df_FilmVote = df_FilmVote.drop(columns = list)



MoH_Vote_Data.head()



# Merge on "person"
MoH_Vote_Data = pd.merge(MoH_Vote_Data , df_FilmVote, on = "person", how='left')


 df_FilmVote.head()

MoH_Vote_Data.columns.tolist()



# Clean Data
list = ['person',
 'name_x',
 'Pure_name',
 'LastName',
 'FirstName',
 'ID_NameParty',
 'first_x',
 'last_x',
 'party_x',
 'pageid_x',
 'wikidataid_x',
 'wikititle_x',
 'name_y',
 'sex_x',
 'ethnicity_x',
 'religion_x',
 'birth_x',
 'death_x',
 'birthplace_x',
 'deathplace_x',
 'Number_Sessions_Served_x',
 'Session_113_x',
 'Session_114_x',
 'Session_115_x',
 'Session_116_x',
 'first_y',
 'last_y',
 'party_y',
 'pageid_y',
 'wikidataid_y',
 'wikititle_y',
 'name_x',



 'death_y',
 'birthplace_y',
 'deathplace_y',
 'Number_Sessions_Served_y',
 'Session_113_y',
 'Session_114_y',
 'Session_115_y',
 'Session_116_y',
 'name_y',
 'party',]

MoH_Vote_Data = MoH_Vote_Data.drop(columns = list)

MoH_Vote_Data.head()

# Save Data-Set externally

compression_opts = dict(method='zip',
                        archive_name='MoH_Vote_AllDonors_Data.csv')  
MoH_Vote_Data.to_csv('MoH_Vote_AllDonors_Data.zip', index=False,
          compression=compression_opts)  

# Convert to Long Format
